+++ src/uni_modules/fant-mini-plus/components/hd-modal/index.ts
@@ -1,13 +1,13 @@
 /*
  * @Author: weisheng
  * @Date: 2022-12-14 17:33:21
- * @LastEditTime: 2023-03-23 13:19:22
+ * @LastEditTime: 2023-05-19 16:16:30
  * @LastEditors: weisheng
  * @Description:
- * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini\components\hd-modal\index.ts
+ * @FilePath: \fant-mini-plus\src\uni_modules\fant-mini-plus\components\hd-modal\index.ts
  * 记得注释
  */
-import { getCurrentInstance } from 'vue'
+import { InjectionKey, Ref, nextTick, provide, ref } from 'vue'
 import { CommonUtil } from '../../index'
 import type { Modal, ModalOptions } from './types'
 
@@ -24,29 +24,33 @@ export const defaultModalOptions: ModalOptions = {
   confirmColor: '#1C64FD'
 }
 
-export function useModal(selector: string = 'hd-modal'): Modal {
-  const { proxy } = getCurrentInstance() as any
+/**
+ * useModal 用到的key
+ *
+ * @internal
+ */
+export const modalDefaultKey = Symbol('__MODAL__') as InjectionKey<Ref<boolean>>
+export const modalDefaultOptionKey = Symbol('__MODAL_OPTION__') as InjectionKey<Ref<ModalOptions>>
 
+export function useModal(selector: string = ''): Modal {
+  const modalShow = ref<boolean>(false) // 是否展示modal
+  const modalOption = ref<ModalOptions>(defaultModalOptions) // Modal选项
+  const modalKey = selector ? '__MODAL__' + selector : modalDefaultKey
+  const modalOptionKey = selector ? '__MODAL_OPTION__' + selector : modalDefaultOptionKey
+  provide(modalOptionKey, modalOption)
+  provide(modalKey, modalShow)
   const showModal = (option: ModalOptions) => {
-    const modal = getModal(proxy, selector)
-    option = CommonUtil.deepMerge(defaultModalOptions, option) as ModalOptions
-    if (modal) {
-      modal.reset(option)
-      modal.show()
+    modalOption.value = CommonUtil.deepMerge(defaultModalOptions, option) as ModalOptions
+    if (modalShow.value) {
+      modalShow.value = false
+      nextTick(() => {
+        modalShow.value = true
+      })
     } else {
-      console.error('未找到 hd-modal 节点，请确认 selector 是否正确')
+      modalShow.value = true
     }
   }
-
   return {
     showModal
-  }
-}
-
-function getModal(proxy, selector: string) {
-  if (proxy && proxy.$refs && proxy.$refs[selector]) {
-    return proxy.$refs[selector]
-  } else {
-    return null
   }
 }