+++ src/uni_modules/fant-mini-plus/components/hd-notify/types.d.ts
@@ -21,17 +21,9 @@ export type NotifyOptions = {
    */
   message: string
   /**
-   * 选择器的选择范围，可以传入自定义组件的 this 作为上下文
-   */
-  context?: any
-  /**
    * 展示时长(ms)，值为 0 时，notify 不会消失，默认值3000
    */
   duration?: number
-  /**
-   * 选择器
-   */
-  selector?: string
   /**
    * 自定义背景色
    */