+++ src/uni_modules/fant-mini-plus/components/hd-toast/index.d.ts
@@ -1,4 +1,12 @@
+import { InjectionKey, Ref } from 'vue'
 import type { Toast, ToastOptions } from './types'
+/**
+ * useToast 用到的key
+ *
+ * @internal
+ */
+export declare const toastDefaultKey: InjectionKey<Ref<boolean>>
+export declare const toastDefaultOptionKey: InjectionKey<Ref<ToastOptions>>
 /**
  * 默认参数
  */